+++ src/main/java/com/sa/bbva/got/model/TipoTramiteCampoKey.java
@@ -14,9 +14,9 @@ public class TipoTramiteCampoKey implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @Column(name = "tipo_tramite_id")
-    Long tipoTramiteId;
+    Integer tipoTramiteId;
 
     @Column(name = "campo_disponible_id")
-    Long campoDisponibleId;
+    Integer campoDisponibleId;
 
 }
